fixed structure migration rollback

+++ src/app/Classes/StructureManager/StructureDestroyer.php
@@ -3,19 +3,26 @@
 namespace LaravelEnso\Core\app\Classes\StructureManager;
 
 use LaravelEnso\Core\app\Models\Menu;
+use LaravelEnso\Core\app\Models\Permission;
 use LaravelEnso\Core\app\Models\PermissionsGroup;
 
 class StructureDestroyer
 {
     private $permissionsGroup;
+    private $permissions;
     private $menu;
 
     public function destroy()
     {
         \DB::transaction(function () {
+            if ($this->permissions->count()) {
+                $this->permissions->each->delete();
+            }
+
             if ($this->permissionsGroup) {
-                $this->permissionsGroup->permissions->each->delete();
-                $this->permissionsGroup->delete();
+                if (!$this->permissionsGroup->permissions->count()) {
+                    $this->permissionsGroup->delete();
+                }
             }
 
             if ($this->menu) {
@@ -26,7 +33,16 @@ class StructureDestroyer
 
     public function setPermissionsGroup($permissionsGroup)
     {
-        $this->permissionsGroup = new PermissionsGroup($permissionsGroup);
+        $this->permissionsGroup = PermissionsGroup::whereName($permissionsGroup['name'])->first();
+    }
+
+    public function setPermissions($permissions)
+    {
+        $this->permissions = collect();
+
+        foreach ($permissions as $permission) {
+            $this->permissions->push(Permission::whereName($permission['name'])->first());
+        }
     }
 
     public function setMenu($menu)
